feat: NodeCollection runtime lifecycle, SceneObjectPtr serialization, Prefab editor support
- Move BeginScene/EndScene/Tick/BeginPlay/EndPlay from Scene to NodeCollection
- World now holds RCPtr<NodeCollection> instead of RCPtr<Scene>
- Prefab can be directly loaded into World for editing
- Add SceneObjectPtr with asset_guid;scene_guid serialization format
- TransformComponent m_parent/m_children changed to SceneObjectPtr
- Extract SceneRuntimeEnvironment to SceneEnvironment.h
- Add virtual GetRuntimeEnvironment() to NodeCollection (Prefab/Scene both override)
- Fix Node SceneObjectGuid auto-generation in SceneObjectConstruct
- Fix Component serialization using SceneObjectGuid instead of ObjectHandle
- Add /bigobj compile option for PulsarEd
- Add Skeleton, AnimationClip assets
- Add skinned mesh GPU skinning pipeline

+++ Pulsar/Source/Pulsar/src/Pulsar/Rendering/PrimitiveStruct.h
@@ -39,7 +39,7 @@ namespace pulsar
         Vector4f Tangent;   // xyz=切线方向，w=副切线符号(+1/-1)
         Color4b  Color;
         Vector2f TexCoords[STATICMESH_MAX_TEXTURE_COORDS];
-        uint32_t BoneIndices[SKINNEDMESH_MAX_BONE_INFLUENCES]; // 骨骼索引
+        uint8_t  BoneIndices[SKINNEDMESH_MAX_BONE_INFLUENCES]; // 骨骼索引（最多256根骨骼）
         float    BoneWeights[SKINNEDMESH_MAX_BONE_INFLUENCES]; // 骨骼权重（归一化，和为1）
     };
     constexpr inline int kSizeofSkinnedMeshVertex = sizeof(SkinnedMeshVertex);
